fix: rename middleware.ts back to proxy.ts — Next.js 16 said so
Turns out Next.js 16 deprecated "middleware" in favor of "proxy".
Our files were originally called proxy.ts and we renamed them
for nothing. The circle of life.
